+++ .gitignore
@@ -1,3 +1,3 @@
 .idea
 vendor
-go-mod-outdated
+go-mod-outdated-library
