+++ outdated/mod/mod_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/psampaz/go-mod-outdated/internal/mod"
+	"github.com/benknigge/go-mod-outdated-library/outdated/mod"
 )
 
 var mods = []mod.Module{
